// src/web/projects/maple-common/src/lib/generated/film-catalog.generated.ts
export type FilmCategory =
  | 'black_white'
  | 'cinema_print'
  | 'color_negative'
  | 'consumer_vintage'
  | 'instant'
  | 'slide';